Rename ExtensionContainerScope to Base
#563
