--- src/Castle.Windsor.Extensions.DependencyInjection/Scope/ExtensionContainerScope.cs
@@ -1,91 +1,0 @@
-// Copyright 2004-2020 Castle Project - http://www.castleproject.org/
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-namespace Castle.Windsor.Extensions.DependencyInjection.Scope
-{
-	using System;
-	using System.Threading;
-
-	using Castle.Core;
-	using Castle.MicroKernel;
-	using Castle.MicroKernel.Lifestyle.Scoped;
-
-	internal class ExtensionContainerScope : ILifetimeScope, IDisposable
-	{
-
-		/// <summary>Current scope for the thread. Initial scope will be set when calling BeginRootScope from a ExtensionContainerRootScope instance.</summary>
-		/// <exception cref="InvalidOperationException">Thrown when there is no scope available.</exception>
-		internal static ExtensionContainerScope Current
-		{
-			get => current.Value ?? throw new InvalidOperationException("No scope available");
-			set => current.Value = value;
-		}
-		protected static readonly AsyncLocal<ExtensionContainerScope> current = new AsyncLocal<ExtensionContainerScope>();
-
-		public static string TransientMarker = "Transient";
-		private readonly ExtensionContainerScope parent;
-		private readonly IScopeCache scopeCache;
-
-		protected ExtensionContainerScope(ExtensionContainerScope parent)
-		{
-			scopeCache = new ScopeCache();
-			current.Value = this;
-			this.parent = parent;
-			RootScope = parent?.RootScope ?? this as ExtensionContainerRootScope;
-		}
-
-		internal ExtensionContainerRootScope RootScope { get; }
-
-		internal static ExtensionContainerScope BeginScope()
-		{
-			return new ExtensionContainerScope(Current);
-		}
-
-		public void Dispose()
-		{
-			if (current.Value == this)
-			{
-				current.Value = parent;
-			}
-
-			if (scopeCache is IDisposable disposableCache)
-			{
-				disposableCache.Dispose();
-			}
-		}
-
-		public Burden GetCachedInstance(ComponentModel model, ScopedInstanceActivationCallback createInstance)
-		{
-			lock (scopeCache)
-			{
-				// Add transient's burden to scope so it gets released
-				if (model.Configuration.Attributes.Get(TransientMarker) == bool.TrueString)
-				{
-					var transientBurden = createInstance((_) => {});
-					scopeCache[transientBurden] = transientBurden;
-					return transientBurden;
-				}
-
-				var scopedBurden = scopeCache[model];
-				if (scopedBurden != null)
-				{
-					return scopedBurden;
-				}
-				scopedBurden = createInstance((_) => {});
-				scopeCache[model] = scopedBurden;
-				return scopedBurden;
-			}
-		}
-	}
-}
